pyproject.toml
[project]
name = "autofin-openfoam"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "numpy-stl",
]
